refactor: move mockDatabase to mock utils
